predictionpipeline/requirements.txt
pandas
sklearn
cassandra-driver
numpy
pyaml